#61 -- Move duplicated methods to BaseTest
